Fix 34 logic bugs across streamIO/copy, real, diffPatch and siblings
Covers all 32 rows recorded in HANDOFF.md for streamIO/copy, streamIO/real,
streamIO/diffPatch, streamIO/fileSystem, streamIO/vector, streamIO/testing,
streamIO/exception, streamIO/adapter and the loose streamIO/*.java files,
plus two undocumented defects found in streamIO/Log.java while working
there: debug() logged at LEVEL_ERROR, and the enter(boolean, String, int)
overload ignored its level parameter, comparing 0 against thresholdLog
instead of level.

Main defect families:
- AOpDouble: eleven long overloads called themselves instead of delegating
  to the double-based op, so any call was a StackOverflowError. They now
  delegate with explicit casts.
- TesterBond.ORat had all four constant-argument branches backwards for OR
  semantics, and the inverse rule set the wrong constant.
- QuantityDouble never assigned its unit parameter in either affected
  constructor, so every unit-dependent call hit an NPE. Note the HANDOFF
  rows name a (double, Unit) and a (double, Unit, double) constructor; the
  actual pair is (ABodyDouble, Unit) and (double, Unit).
- RandomGauss/RandomGauss2 compared the cached deviate against itself, so
  the cached second value was never returned.
- FilterInMul.getMinDouble returned a wrong-signed bound for a negative
  scale; StreamIn_Geometric assigned its fields out of order.
- Fraction/FractionLong Floor() divided the wrong way round.
- VersionTree.LESS never returned false for a greater element; writeTo
  resized after writing and dropped the resized array.
- ACopyAble.toStream recursed forever on a cyclic object graph; it now has
  a ThreadLocal re-entrancy guard that emits ClassName@identityHash.
- AReSetAble.JUMP ignored negative offsets; Log's shared DateFormat was
  used unsynchronized.

Two changes rest on judgement rather than an unambiguous specification and
want review: AShiftAble's carry threading, which the author had left
explicitly undefined and which now follows the in/out contract documented
on IShiftAble; and the ACopyAble cycle guard's output format.

All 27 touched files compile under javac 11 and keep CRLF endings.
Verification is compile-only - there is no test suite for these packages.

+++ streamIO/copy/primitiveOp/AOpDouble.java
@@ -48,52 +48,41 @@ implements IOpDouble {
 		return ((IOpDouble) copy(Integer.MAX_VALUE)).divAt(arg); }
 
 	/**Copy in Place: =	*/
-	// TODO: LOGIC: infinite recursion - calls itself with the same signature/args instead of delegating to a double-based operation; any caller of copyAt(long) triggers a StackOverflowError.
-	public IOpLong copyAt(long arg) { return copyAt(arg); }
+	public IOpLong copyAt(long arg) { return copyAt((double) arg); }
 
 	/**Check for equality: ==	*/
-	// TODO: LOGIC: infinite recursion - calls itself with the same signature/args; any caller of equals(long) triggers a StackOverflowError.
-	public boolean equals(long arg) { return equals(arg); }
+	public boolean equals(long arg) { return equals((double) arg); }
 
 	/**Comparison with a long Number: < arg	 */
-	// TODO: LOGIC: infinite recursion - calls itself with the same signature/args; any caller of less(long) triggers a StackOverflowError.
-	public boolean less(long arg) { return less(arg); }
+	public boolean less(long arg) { return less((double) arg); }
 
 	/**Comparison with a long Number: > arg	 */
-	// TODO: LOGIC: infinite recursion - calls itself with the same signature/args; any caller of grtr(long) triggers a StackOverflowError.
-	public boolean grtr(long arg) { return grtr(arg); }
+	public boolean grtr(long arg) { return grtr((double) arg); }
 
 	/**Maximum in Place: 	*/
-	// TODO: LOGIC: infinite recursion - calls itself with the same signature/args; any caller of MaxAt(long) triggers a StackOverflowError.
-	public IOpLong MaxAt (long arg) { return MaxAt(arg); }
+	public IOpLong MaxAt (long arg) { return MaxAt((double) arg); }
 
 	/**Minimum in Place: 	*/
-	// TODO: LOGIC: infinite recursion - calls itself with the same signature/args; any caller of MinAt(long) triggers a StackOverflowError.
-	public IOpLong MinAt (long arg) { return MinAt(arg); }
+	public IOpLong MinAt (long arg) { return MinAt((double) arg); }
 
 	/**Addition of a long Number in Place: += arg	 */
-	// TODO: LOGIC: infinite recursion - calls itself with the same signature/args; any caller of addAt(long) triggers a StackOverflowError.
-	public IOpLong addAt(long arg) { return addAt(arg); }
+	public IOpLong addAt(long arg) { return addAt((double) arg); }
 
 	/**Subtraction of a long Number in Place: -= arg	 */
-	// TODO: LOGIC: infinite recursion - calls itself with the same signature/args; any caller of subAt(long) triggers a StackOverflowError.
-	public IOpLong subAt(long arg) { return subAt(arg); }
+	public IOpLong subAt(long arg) { return subAt((double) arg); }
 
 	/**Multiplication of a long Number in Place: *= arg	 */
-	// TODO: LOGIC: infinite recursion - calls itself with the same signature/args; any caller of mulAt(long) triggers a StackOverflowError.
-	public IOpLong mulAt(long arg) { return mulAt(arg); }
+	public IOpLong mulAt(long arg) { return mulAt((double) arg); }
 
 	/**Division of a long Number in Place: /= arg	 */
-	// TODO: LOGIC: infinite recursion - calls itself with the same signature/args; any caller of divAt(long) triggers a StackOverflowError.
-	public IOpLong divAt(long arg) { return divAt(arg); }
+	public IOpLong divAt(long arg) { return divAt((double) arg); }
 
 	/**  Linear Mapping: x * a + y			*/
 	public IOpDouble Lin		(double a, double y) {
 		return ((IOpDouble) copy(Integer.MAX_VALUE)).LinAt(a, y); }
 
 	/**  Linear Mapping in Place: x*=a + y	<=> x*=a; x+=y; */
-	// TODO: LOGIC: infinite recursion - calls itself with the same signature/args instead of delegating to the long-based ops (mulAt/addAt); any caller triggers a StackOverflowError.
 	public IOpLong LinAt		(long a, long y) {
-		return LinAt(a, y); }
+		return LinAt((double) a, (double) y); }
 
 }
